Admin portal with users, categories completed + top down schema rewrite + react hooks rewrite

+++ server/controllers/CategoryController.js
@@ -1,24 +1,44 @@
 const Category = require('../models/CategorySchema');
 
-// Create a category
-exports.create = (req, res) => {
-  const category = new Category(req.body);
-
-  category.save((err) => {
-    if (err) {
-      console.log(err);
-      res.status(400).send(err);
-    } else {
-      res.json(category);
-      console.log(category);
-    }
-  });
+// Get all the categories
+/* 
+  Accepts query parameters in this format
+  children=true // or false
+  providers=true // or false
+  True will populate the array, false will leave it as an array of ObjectIDs.
+*/
+exports.list = (req, res) => {
+  let populateOptions = [];
+  if (req.query.children === 'true')
+    populateOptions = [...populateOptions, 'children'];
+  if (req.query.providers === 'true')
+    populateOptions = [...populateOptions, 'providers'];
+  Category.find({})
+    .populate(...populateOptions)
+    .exec((err, categories) => {
+      if (err) {
+        res.status(400).send(err);
+      } else {
+        res.json(categories);
+      }
+    });
 };
 
-// Get all the categories
-exports.list = (req, res) => {
-  Category.find({})
-    .populate('subcategory_of')
+// Get the top level categories
+/* 
+  Accepts query parameters in this format
+  children=true // or false
+  providers=true // or false
+  True will populate the array, false will leave it as an array of ObjectIDs.
+*/
+exports.listTopLevel = (req, res) => {
+  let populateOptions = [];
+  if (req.query.children === 'true')
+    populateOptions = [...populateOptions, 'children'];
+  if (req.query.providers === 'true')
+    populateOptions = [...populateOptions, 'providers'];
+  Category.find({ is_subcategory: false })
+    .populate(...populateOptions)
     .exec((err, categories) => {
       if (err) {
         console.log(err);
@@ -29,37 +49,29 @@ exports.list = (req, res) => {
     });
 };
 
-// Get the top level categories
-exports.listTopLevel = (req, res) => {
-  Category.find({ subcategory_of: { $size: 0 } }).exec((err, categories) => {
-    if (err) {
-      console.log(err);
-      res.status(400).send(err);
-    } else {
-      res.json(categories);
-    }
-  });
-};
-
-// Get the subcategories
-exports.listSubCategory = (req, res) => {
-  Category.find({ 'subcategory_of.0': req.id }).exec((err, categories) => {
-    if (err) {
-      console.log(err);
-      res.status(400).send(err);
-    } else {
-      res.json(categories);
-    }
-  });
-};
-
 // Get the current category
 exports.read = (req, res) => {
   res.json(req.category);
 };
 
+// Create a category
+exports.create = (req, res) => {
+  if (req.user.role !== 'Owner') return res.status(403).end();
+  const category = new Category(req.body);
+
+  category.save((err) => {
+    if (err) {
+      console.log(err);
+      res.status(400).send(err);
+    } else {
+      res.json({ success: true, category: category });
+    }
+  });
+};
+
 // Update a category
 exports.update = (req, res) => {
+  if (req.user.role !== 'Owner') return res.status(403).end();
   const category = req.category;
   const newCategory = req.body;
 
@@ -74,26 +86,39 @@ exports.update = (req, res) => {
       console.log(err);
       res.status(400).send(err);
     } else {
-      res.json(category);
-      console.log(category);
+      res.json({ success: true, category: category });
     }
   });
 };
 
 exports.delete = (req, res) => {
+  if (req.user.role !== 'Owner') return res.status(403).end();
   const category = req.category;
   Category.deleteOne(category, (err) => {
     if (err) {
       console.log(err);
       res.status(400).send(err);
-    } else res.status(200).end();
+    } else res.json({ success: true });
   });
 };
 
 // Middleware to get a category from database by ID, save in req.category
+/* 
+  Accepts query parameters in this format
+  children=true // or false
+  providers=true // or false
+  True will populate the array, false will leave it as an array of ObjectIDs.
+*/
 exports.categoryById = (req, res, next, id) => {
+  let populateOptions = [];
+  console.log(req.query);
+  if (req.query.children === 'true')
+    populateOptions = [...populateOptions, 'children'];
+  if (req.query.providers === 'true')
+    populateOptions = [...populateOptions, 'providers'];
+  console.log(populateOptions);
   Category.findById(id)
-    .populate('subcategory_of')
+    .populate(...populateOptions)
     .exec((err, category) => {
       if (err) {
         res.status(400).send(err);
