+++ server/controllers/ProviderController.js
@@ -2,6 +2,7 @@ const Provider = require('../models/ProviderSchema');
 
 // Create a provider
 exports.create = (req, res) => {
+  if (req.user.role !== 'Owner') return res.status(403).end();
   const provider = new Provider(req.body);
 
   provider.save((err) => {
@@ -9,8 +10,7 @@ exports.create = (req, res) => {
       console.log(err);
       res.status(400).send(err);
     } else {
-      res.json(provider);
-      console.log(provider);
+      res.json({ success: true, provider: provider });
     }
   });
 };
@@ -22,6 +22,7 @@ exports.read = (req, res) => {
 
 // Update a provider
 exports.update = (req, res) => {
+  if (req.user.role !== 'Owner') return res.status(403).end();
   // TODO: UPDATE FIELDS FROM BODY?
   const provider = req.provider;
 
@@ -30,67 +31,44 @@ exports.update = (req, res) => {
       console.log(err);
       res.status(400).send(err);
     } else {
-      res.json(provider);
-      console.log(provider);
+      res.json({ success: true, provider: provider });
     }
   });
 };
 
 // Delete a provider
 exports.delete = (req, res) => {
+  if (req.user.role !== 'Owner') return res.status(403).end();
   const provider = req.provider;
   Provider.deleteOne(provider, (err) => {
     if (err) {
       console.log(err);
       res.status(400).send(err);
-    } else res.status(200).end();
+    } else res.json({ success: true });
   });
 };
 
 // Get all the providers
 exports.list = (req, res) => {
-  Provider.find({})
-    .populate('categories')
-    .exec((err, providers) => {
-      if (err) {
-        console.log(err);
-        res.status(400).send(err);
-      } else {
-        res.json(providers);
-      }
-    });
-};
-
-// Get all the providers in a given subcategory
-exports.listSubCategory = (req, res) => {
-  res.json(req.inSubcategory);
+  Provider.find({}).exec((err, providers) => {
+    if (err) {
+      console.log(err);
+      res.status(400).send(err);
+    } else {
+      res.json(providers);
+    }
+  });
 };
 
 // Middleware to get a provider from database by ID, save in req.provider
 exports.providerById = (req, res, next, id) => {
-  Provider.findById(id)
-    .populate('categories')
-    .exec((err, provider) => {
-      if (err) {
-        res.status(400).send(err);
-      } else {
-        req.provider = provider;
-        req.id = id;
-        next();
-      }
-    });
-};
-
-// Middleware to get providers in subcategory, save in req.inSubcategory
-exports.getSubCategory = (req, res, next, id) => {
-  Provider.find({ categories: id })
-    .populate('categories')
-    .exec((err, providers) => {
-      if (err) {
-        res.status(400).send(err);
-      } else {
-        req.inSubcategory = providers;
-        next();
-      }
-    });
+  Provider.findById(id).exec((err, provider) => {
+    if (err) {
+      res.status(400).send(err);
+    } else {
+      req.provider = provider;
+      req.id = id;
+      next();
+    }
+  });
 };
